+++ packages/atlas/src/joystream-lib/errors.ts
@@ -33,7 +33,8 @@ export enum ErrorCode {
   ChannelStateBloatBondChanged = 'ChannelStateBloatBondChanged', // video state bloat bond changed
   InsufficientBalance = 'InsufficientBalance', // balance to low to send tokens
   InsufficientBalanceForChannelCreation = 'InsufficientBalanceForChannelCreation', // Cannot create the channel: channel creator has insufficient balance, (budget for channel state bloat bond + channel data objs state bloat bonds + data objs storage fees))
-  InsufficientBalanceForVideoCreation = ' InsufficientBalanceForVideoCreation,', /// Cannot create the video: video creator has insufficient balance,  (budget for video state bloat bond + video data objs state bloat bonds + data objs storage fees)
+  InsufficientBalanceForVideoCreation = 'InsufficientBalanceForVideoCreation,', /// Cannot create the video: video creator has insufficient balance,  (budget for video state bloat bond + video data objs state bloat bonds + data objs storage fees)
+  InsufficientBalanceToCoverPayment = 'InsufficientBalanceToCoverPayment', // Insufficient balance to cover MemberRemark payment
   TokenDoesNotExist = 'TokenDoesNotExist', // Requested token does not exist
   NoActiveSale = 'NoActiveSale', // The token has no active sale at the moment
   InsufficientBalanceForTokenPurchase = 'InsufficientBalanceForTokenPurchase', // Account's JOY balance is insufficient to make the token purchase