+++ CHANGELOG.md
@@ -5,6 +5,18 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [5.6.0] - 2024-12-16
+
+**IMPORTANT:** Depends on Orion release `4.2.0`.
+
+### Added
+
+- Video tipping functionality (https://github.com/Joystream/atlas/issues/6291)
+
+### Fixed
+
+- Improved error handling for optimistic transactions (handle `ExtrinsicFailed` errors)
+
 ## [5.5.0] - 2024-11-07
 
 **IMPORTANT:** Depends on Orion release `4.1.0`.